+++ CHANGELOG.md
@@ -10,6 +10,8 @@
     -   Changed `Clear Mod` to `Reset` in the sheet.
     -   Allow the clicking on a bot in the sheet in single selection mode to deselect the bot.
     -   Changed `onCombine()` action tag to `onCombine(#tag:"value")` and set the autofill to not auto add this tag to the sheet.
+-   Bug Fixes
+    -   Dragging a bot with no bot selected will no longer select a mod of the dragged bot.
 
 ## V0.9.22